+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## Gazebo 10.x.x (201x-xx-xx)
 
+1. Dart heightmaps with bullet and ODE collision detectors
+    * [Pull request #2956](https://bitbucket.org/osrf/gazebo/pull-request/2956)
+    * [Pull request #3066](https://bitbucket.org/osrf/gazebo/pull-request/3066)
+
 1. Add record\_resources field to gazebo log msgs
     * [Pull request #2797](https://bitbucket.org/osrf/gazebo/pull-request/2797)
     * [Pull request #3008](https://bitbucket.org/osrf/gazebo/pull-request/3008)